moved getOptionValues to a util file
