+++ tsconfig.app.json
@@ -30,10 +30,9 @@
     "baseUrl": "./src", // This is to avoid adding the ./src prefix to each path below.
     "paths": {
       "@components/*": ["components/*"],
-      "@types": ["types/index.ts"]
-      // "@hooks/*": ["hooks/*"],
-      // "@utils/*": ["utils/*"],
-      // "@config/*": ["config/*"]
+      "@types": ["types/index.ts"],
+      "@hooks/*": ["hooks/*"],
+      "@utils/*": ["utils/*"]
     }
   },
   "include": ["src"]
